Refractor webhookPublish function
move the logic of sending HTTP POST request to webhook
URL to a new function.
later the same function can be used for testing
connection between peers and webhook URL.
